Fix TopOrder implementation
DFS in TopOrder was wrong, because it pushed all children to stack at
once. Instead, each child should be able to push its own children first
(to be a proper DFS), so keep track of which child is next in the stack
as well.

Doing a proper DFS with manual stack requires keeping child iteration
progress in stack. The regular RgfaGraph iterators would have worked,
but they also keep a pointer to the Graph, so this commit also adds an
API to query edges by edge_id and advancing to next edge_id manually.

This way only edge_id can be stored in stack, saving some memory.