Add ComplexityEstimator
ComplexityEstimator estimates an upper bound on the number of kmers at
the start and end of each node. These can be used as hints to run
a different kmer computation algorithm on more complex parts of the
graph.

Also add ce-test mode to main, which checks if the estimated values are
greater-or-equal to actual values produced by a full triegraph builder.