Enable easier switching TrieData base datastructure
TrieData consists of 2 maps, trie2graph and graph2trie, and an
additional bitset that keeps which (shorter) kmers are in the trie and
which aren't.

This commit abstracts away the 2 maps in a base class, so the core data
structures can be changed, without affecting the higher level API of
TrieData.

Currently DenseMultimap and SimpleMultimap are supported.