Implement HybridMultimap
This is another one in the multimap family for use in the TrieData
class. It combines a regular map with a signle vector for all values.

The advantage over an in-library multimap is that out-of-std maps
(google sparce_hash_map and sparcepp) lack multimap, so this can be used
as an adapter to make any map into a multimap.