Implement zero-mem-overhead TrieData building
Currently the code is controlled from a template boolean, because for it
to work:
- VectorPairs has to be DUAL
- sizeof of the two pair types need to match (LetterLoc, KHolder)
- DenseMultimap with SortedVector used for both t2g and g2t

Otherwise the algorithm doesn't make much sense.

Possibly move the boolean to Manager, checking if all the prereq are met
and additionally (possibly) enable flag (at the moment the zero-mem-overhead
is slower, so it should be used only when necessary).