Introduce CfgFlags, rewrite tests
To enable easier configuration of the config (DnaConfig at the moment),
introduce CfgFlags which are just a bunch of booleans stuck in a u32.

Add testlib/dna.h which includes the most used test configuration, so it
doesn't have to be redeclared every time.

Also, add a flag for the zero-overhead TrieData building, because even
if all preconditions are met (i.e we *can* do zero-overhead TrieData
from VectorPairs), it might not be desired, because it's much slower.

Also at the moment not all preconditions can be easily checked in
template code, so adding a flag makes it the user responsibility if
zero-overhead can be done safely.