Add CompactVector with tests
CompactVector is like a regular vector, but it uses only a given number
of bits to store each value (instead of the full value-type (32, 64
bits)). Thus each operation requires some bit-manipulations to set/get
the value at a given index.

It is apporx 3x slower than a regular vector (and up to 2x less memory).