Implement random-access-iterator for SortedVector
One drawback of the current implementation is that all operations check
if the index is one-after-the-end, so they can apply 0 diff to move to
size()-1 from end()

An alternative solution would be to have .diffs store size()+1 elements
(the last one being zero), but due to the way SortedVector is
initialized (.reserve() + many .push_back()s), this approach would
require non-trivial logic in push_back to ensure that there is always
a zero after the end.

Another alternative is to have forward-only iterator, but it still has
to check if it transitions from last to one-after-last.