Switch VectorPairs impl to multiple classes
Previously the 3 types of VectorPairs could be instantiated by passing
the VectorPairsImpl as a third template argument. This is a bit more
convenient, but it prevents specializations from defining additional
template arguments.

VectorPairs DUAL should be able to deal with CompactVector, so this
switch was done in preparation for that.