Use reference instead of reference_type
Iterators and containers define nested types for easier template
programming. It turns out reference is the correct name for the
reference type (and not reference_type, which looks like value_type
which is correct).

NOTE: God knows how many weird complier errors were caused by this.
We'll never know. It would be nice if iterator concepts would catch this
:/